requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
requests==2.31.0
Pillow==10.0.1
python-socketio==5.8.0
python-engineio==4.7.1
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
dnspython==2.4.2
idna==3.4
urllib3==2.0.7
certifi==2023.7.22
charset-normalizer==3.2.0 